requirements.txt
python-telegram-bot==21.3
openai==1.30.1
motor==3.6.0
redis==5.0.4
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
